Move Sidebar to components/core
