+++ app/frontend-webapp/app/layout.tsx
@@ -5,7 +5,7 @@ import { Toaster } from "@/components/shadow-ui/Toast/Toaster";
 import { Providers } from "@/app/providers";
 import { css } from "@/styled-system/css";
 import { Metadata } from "next";
-import { Sidebar } from "@/components/sidebar/Sidebar";
+import { Sidebar } from "@/components/core/sidebar/Sidebar";
 import { SidebarItems } from "@/config/sidebar";
 import Navbar from "@/components/Navbar";
