+++ app/frontend-webapp/components/core/sidebar/SidebarContent.tsx
